feat: add path alias and extract particle class into seperate file
