+++ src/flows/audioVisualizer/index.ts
@@ -1,1 +1,2 @@
 export * as default from "./audioVisualizer";
+export * from "./Particle";
