+++ vite.config.js
@@ -1,5 +1,6 @@
 import { defineConfig } from "vite";
 import inject from "@rollup/plugin-inject";
+import path from "path";
 
 export default defineConfig({
   plugins: [
@@ -7,4 +8,9 @@ export default defineConfig({
       p5: "p5",
     }),
   ],
+  resolve: {
+    alias: {
+      src: path.resolve(__dirname, "./src"),
+    },
+  },
 });
